update to use js-cookie not chuangbo:cookie which has bugs
chuangbo:cookie  does not correctly set cookies - instead of fixing it, I moved to js-cookies.

It appears that this package was trying to set the intended cookie in it's entirety with document.cookie, when in fact document.cookie needs to be called each time you wish to add an additional cookie.

https://stackoverflow.com/questions/26349052/why-would-setting-document-cookie-not-work-in-chrome